+++ components/o-forms/src/js/error-summary.js
@@ -13,21 +13,23 @@ class ErrorSummary {
 	 * Class constructor.
 	 *
 	 * @param {Array<ErrorSummaryElement>} [elements] - An array of objects, where each object describes an invalid input element
+	 * @param {string} [headingMessage='There is a problem'] - A message to show in the header. It defaults to: 'There is a problem'
 	 * @example
-	 * const example = [
-	 *	{
-	 *		id: 'text-input',
-	 *		valid: false,
-	 *		error: 'Please fill out this field'
-	 *		label: 'Input Label',
-	 *		element: <Element>
-	 *	}
-	 *	...
-	 *	]
-	 *	new ErrorSummary(example)
+	 *	const elementsExample = [
+	 *		{
+	 *			id: 'text-input',
+	 *			valid: false,
+	 *			error: 'Please fill out this field',
+	 *			label: 'Input Label',
+	 *			element: <Element />,
+	 *		},
+	 *		{...},
+	 *	];
+	 *	new ErrorSummary(example, 'This is a heading message')
 	 */
-	constructor(elements) {
+	constructor(elements, headingMessage) {
 		this.elements = elements;
+		this.headingMessage = headingMessage || 'There is a problem';
 		const hasAnInverseField = elements.some(elem => {
 			if (elem.field) {
 				return elem.field.classList.contains('o-forms-field--inverse');
@@ -56,8 +58,8 @@ class ErrorSummary {
 		}
 		div.setAttribute('aria-labelledby', 'error-summary');
 		div.setAttribute('role', 'alert');
-		div.innerHTML = '<h4 class="o-forms__error-summary__heading" id="error-summary">There is a problem</h4>';
 
+		div.innerHTML = `<h4 class="o-forms__error-summary__heading" id="error-summary">${this.headingMessage}</h4>`;
 		div.appendChild(ErrorSummary.createList(invalidInputs));
 		return div;
 	}
@@ -85,9 +87,12 @@ class ErrorSummary {
 			// invalid input but with no label to create an error summary
 			if (input.valid === false && !input.label) {
 				// eslint-disable-next-line no-console
-				console.warn(`Could not add an invalid input to the error summary. ` +
-				`Check the input has a parent \`o-forms-field\` element with correct title markup. ` +
-				`Or disable the error summary feature for this form with \`data-o-forms-error-summary="false"\`.`, input.element);
+				console.warn(
+					`Could not add an invalid input to the error summary. ` +
+						`Check the input has a parent \`o-forms-field\` element with correct title markup. ` +
+						`Or disable the error summary feature for this form with \`data-o-forms-error-summary="false"\`.`,
+					input.element
+				);
 			}
 			// invalid input, add to error summary
 			if (input.valid === false && input.label) {
@@ -117,8 +122,11 @@ class ErrorSummary {
 		}
 		// If no id exist return an error summary item without a link.
 		// eslint-disable-next-line no-console
-		console.warn(`Could not link to an invalid input from the error summary. ` +
-			`Add a unique id attribute to the input element.`, input.element);
+		console.warn(
+			`Could not link to an invalid input from the error summary. ` +
+				`Add a unique id attribute to the input element.`,
+			input.element
+		);
 
 		item.innerHTML = ErrorSummary._getItemContent(input);
 		return item;
@@ -133,10 +141,13 @@ class ErrorSummary {
 	static createAnchor(input) {
 		const anchor = document.createElement('a');
 		anchor.setAttribute('href', `#${input.id}`);
-		anchor.addEventListener('click', function(e) {
-			e.preventDefault();
-			document.getElementById(this.id).focus();
-		}.bind(input));
+		anchor.addEventListener(
+			'click',
+			function (e) {
+				e.preventDefault();
+				document.getElementById(this.id).focus();
+			}.bind(input)
+		);
 		anchor.innerHTML = ErrorSummary._getItemContent(input);
 		return anchor;
 	}
@@ -147,9 +158,11 @@ class ErrorSummary {
 	 * @returns {string} - the html text for an error summary item
 	 */
 	static _getItemContent(input) {
-		return '<span class="o-forms__error-summary__item-overview">' +
+		return (
+			'<span class="o-forms__error-summary__item-overview">' +
 			`${input.label}</span>: ` +
-			`<span class="o-forms__error-summary__item-detail">${input.error}</span>`;
+			`<span class="o-forms__error-summary__item-detail">${input.error}</span>`
+		);
 	}
 }