+++ components/o-message/stories/internal/message-action.stories.tsx
@@ -2,13 +2,11 @@ import withHtml from 'origami-storybook-addon-html';
 import {withDesign} from 'storybook-addon-designs';
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 import {useEffect} from 'react';
-import {ActionMessage} from '../src/tsx/message';
-import javascript from '../main';
+import {ActionMessage} from '../../src/tsx/message';
+import javascript from '../../main';
 import './message.scss';
 
-const Brand = process.env.ORIGAMI_STORYBOOK_BRAND || 'core';
-
-const actionProps = {
+export default {
 	title: 'Components/o-message',
 	component: ActionMessage,
 	decorators: [withDesign, withHtml],
@@ -35,10 +33,6 @@ const actionProps = {
 		},
 	},
 };
-
-
-const defaultProps = Brand !== 'core' ? actionProps : null;
-export default defaultProps as ComponentMeta<typeof ActionMessage>;
 
 const ActionStory = args => {
 	useEffect(() => {